requirements.txt
asgiref==3.3.4
attrs==21.2.0
Django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
importlib-metadata==4.5.0
iniconfig==1.1.1
networkx==2.5.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
psycopg2-binary==2.8.4
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1
